Fixed deleting issues

+++ backend/routers/FeaturesCnnRouter.py
@@ -38,13 +38,21 @@ def add_feature_cnn(feature_cnn_body: FeatureCnnSummary):
 
 @router.delete("/{feature_cnn_id}", status_code=status.HTTP_200_OK)
 def delete_feature_cnn(feature_cnn_id: int):
-    try:
-        cursor.execute("delete from features_cnn where id = ?", (feature_cnn_id,))
-        logging.info(f"feature cnn with id = {feature_cnn_id} has been deleted successfully")
-    except mariadb.Error as e:
-        logging.error(f"Could not delete feature cnn with id = {str(feature_cnn_id)}. Error: {e}")
-        return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
-                            content=f"Could not delete feature cnn with id = {str(feature_cnn_id)}")
+    get_response = get_feature_cnn(feature_cnn_id)
+    if get_response.status_code == status.HTTP_200_OK:
+        try:
+            cursor.execute("delete from features_cnn where id = ?", (feature_cnn_id,))
+            connection.commit()
+            logging.info(f"feature cnn with id = {feature_cnn_id} has been deleted successfully")
+        except mariadb.Error as e:
+            logging.error(f"Could not delete feature cnn with id = {str(feature_cnn_id)}. Error: {e}")
+            return JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                                content=f"Could not delete feature cnn with id = {str(feature_cnn_id)}")
+    else:
+        logging.error(f"Could not delete feature cnn with id = {str(feature_cnn_id)}. Error: entity does not exist.")
+        JSONResponse(status_code=status.HTTP_400_BAD_REQUEST,
+                     content=f"Could not delete feature cnn with id = {str(feature_cnn_id)} because entity does not "
+                             f"exist.")
 
 
 @router.get("/", status_code=status.HTTP_200_OK)
